% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_sampler_exist}
\alias{check_sampler_exist}
\title{Check of the sampler's existence in the catalog.}
\usage{
check_sampler_exist(sampler_name = NULL)
}
\arguments{
\item{sampler_name}{description}
}
\value{
description
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Checks if a sampler already exists in the catalog or not.
}
\examples{
\dontrun{
## Not run:
check_sampler_exist(sampler_name = "Niskin bottle")

}
## End (Not run)
}
\author{
Alessandro Oggioni, phD (2023) \email{oggioni.a@cnr.it}
}
