% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{specimen_catalogue}
\alias{specimen_catalogue}
\title{Create a catalogue of samples}
\usage{
specimen_catalogue(excel_path = NULL)
}
\arguments{
\item{excel_path}{A \code{string} of the location where the spreadsheet template
is located}
}
\value{
A folder in the working directory containing two files, TTL and XML,
for each sample listed in the 'SpecimenInfo' worksheet of the spreadsheet
template.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function generates a series of files in TTL and XML formats based on a
spreadsheet template ('specimen_template.xlsx'). These files reproduce the
contents of each row according to the Semantic Sensor Network Ontology (SSN)
model and the relevant standards. For more details on compatibility, refer
to: https://github.com/oggioniale/specimen_catalogue?tab=readme-ov-file.
}
\examples{
\dontrun{
## Not run:
specimen_catalogue(excel_path = "./specimen_template.xlsx")

}
## End (Not run)

}
\author{
Alessandro Oggioni, phD (2023) \email{oggioni.a@cnr.it}
}
