% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{specimen_ttl}
\alias{specimen_ttl}
\title{Create the ttl of sample}
\usage{
specimen_ttl(
  excel_file = NULL,
  excel_curators = NULL,
  excel_relation = NULL,
  excel_sampler = NULL,
  specimen_uuids = NULL,
  root_dir = NULL,
  creator_name = creator_name,
  creator_surname = creator_surname,
  creator_orcid = creator_orcid
)
}
\arguments{
\item{excel_file}{A \code{tibble} that contains the columns present in the
'SpecimenInfo' sheet of the spreadsheet template.}

\item{excel_curators}{A \code{tibble} that contains the columns present in the
'CuratorsInfo' sheet of the spreadsheet template.}

\item{excel_relation}{A \code{tibble} that contains the columns present in the
'RelationInfo' sheet of the spreadsheet template.}

\item{excel_sampler}{A \code{tibble} that contains the columns present in the
'SamplersInfo' sheet of the spreadsheet template.}

\item{specimen_uuids}{A \code{character} that contains a number of strings
UUIDs equal to the number of samples present in the SpecimenInfo' sheet
of the spreadsheet template.}

\item{root_dir}{A \code{string} with the path to the folder where the ttl and
XML files will be written.}
}
\value{
A list of tibbles, each one being the validation result for a
single TTL file, as returned by \code{\link[=specimen_validate_ttl]{specimen_validate_ttl()}}.
}
\description{
This function generates Turtle (TTL) files for each sample listed
in the input \code{excel_file}, using SOSA/SSN, PROV-O, DCAT and related
vocabularies. For each generated TTL, a SHACL-like validation is
performed via \code{\link[=specimen_validate_ttl]{specimen_validate_ttl()}}, and a per-file validation
tibble is returned.
}
\author{
Alessandro Oggioni, phD (2023) \email{oggioni.a@cnr.it}
}
\keyword{internal}
