% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speciment-package.R
\docType{package}
\name{Speciment-package}
\alias{Speciment}
\alias{Speciment-package}
\title{Speciment: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oggioniale/specimen_catalogue}
  \item \url{https://oggioniale.github.io/specimen_catalogue/}
  \item Report bugs at \url{https://github.com/oggioniale/specimen_catalogue/issues}
}

}
\author{
\strong{Maintainer}: Alessandro Oggioni \email{alessandro.oggioni@cnr.it} (\href{https://orcid.org/0000-0002-7997-219X}{ORCID}) [funder]

}
\keyword{internal}
