% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{specimen_runApp}
\alias{specimen_runApp}
\title{Run shiny app for get and visualize sample catalogue}
\usage{
specimen_runApp(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function run the Shiny App
}
\examples{
\dontrun{
## Not run:
specimen_runApp(launch.browser = rstudioapi::viewer)

}
## End (Not run)

}
\author{
Alessandro Oggioni, phD \email{alessandro.oggioni@cnr.it}
}
